# [chore] optimize graph add node operation