## Description

<!-- Add a more detailed description of the changes if needed. -->

This PR optimizes query graph's adding node operation by removing redundant node serialization & deserialization steps.

## Performance Comparison

Script used (feature list contains 204 features)
```
from datetime import datetime

import featurebyte as fb
from featurebyte.common.model_util import get_version
from featurebyte.models.base import VersionIdentifier
from featurebyte.models.feature_list import FeatureListModel
from featurebyte.schema.feature import BatchFeatureCreate, FeatureCreate

fb.list_catalogs()
catalog = fb.activate_and_get_catalog("Grocery V1 - playground (spark)")

feature_list_id = catalog.list_feature_lists()["id"].iloc[0]
feature_list = catalog.get_feature_list_by_id(feature_list_id)

def profile_runtime(feature_list: FeatureListModel):
    # start profiling time
    start, next_start = datetime.now(), datetime.now()
    feature_payloads = []
    for feat in feature_list.feature_objects.values():
        feature_payloads.append(FeatureCreate(**feat.dict(by_alias=True)))
    start, next_start = next_start, datetime.now()
    print(f"Time taken to create feature payloads: {next_start - start}")

    features = []
    for feat in feature_list.feature_objects.values():
        features.append(feat.cached_model)
    start, next_start = next_start, datetime.now()
    print(f"Time taken to create feature models: {next_start - start}")

    feature_list_create = BatchFeatureCreate.create(
        features=feature_payloads,
    )
    start, next_start = next_start, datetime.now()
    print(f"Time taken to create feature list create payload: {next_start - start}")

    for _ in feature_list_create.iterate_features():
        pass
    start, next_start = next_start, datetime.now()
    print(f"Time taken to iterate create feature list create payload: {next_start - start}")

    _ = FeatureListModel(
        name="feature_list",
        version=VersionIdentifier(name=get_version()),
        feature_ids=[feat.id for feat in features],
        features=features,
    )
    start, next_start = next_start, datetime.now()
    print(f"Time taken to create feature list model: {next_start - start}")


def profile_create_save_and_delete(feature_list):
    start, next_start = datetime.now(), datetime.now()
    new_feature_list = fb.FeatureList(list(feature_list.feature_objects.values()), name="testing")
    start, next_start = next_start, datetime.now()
    print(f"Time taken to create feature list: {next_start - start}")

    new_feature_list.save()
    start, next_start = next_start, datetime.now()
    print(f"Time taken to save feature list: {next_start - start}")

    new_feature_list.delete()
    start, next_start = next_start, datetime.now()
    print(f"Time taken to delete feature list: {next_start - start}")


profile_runtime(feature_list)
profile_create_save_and_delete(feature_list)
```

main:
```
(featurebyte-py3.8) ➜  featurebyte git:(main) ✗ python opt_feature_list.py
2023-06-14 07:53:32 | INFO     | featurebyte.api.catalog | activate:157 | Catalog activated: Grocery V1 - playground (spark)
Loading Feature(s) |████████████████████████████████████████| 204/204 [100%] in 1.1s (173.01/s) 
Time taken to create feature payloads: 0:00:01.380866
Time taken to create feature models: 0:00:02.985411
Time taken to create feature list create payload: 0:00:00.362690
Time taken to iterate create feature list create payload: 0:00:01.015061
Time taken to create feature list model: 0:00:02.051015
Time taken to create feature list: 0:00:00.001926
Done! |████████████████████████████████████████| 100% in 0.1s (9.88%/s) 
Loading Feature(s) |████████████████████████████████████████| 204/204 [100%] in 1.0s (181.68/s) 
Time taken to save feature list: 0:00:07.775290
Time taken to delete feature list: 0:00:01.031623
```

current branch:
```
(featurebyte-py3.8) ➜  featurebyte git:(senbong/optimize-graph-add-node) ✗ python opt_feature_list.py
2023-06-14 07:49:46 | INFO     | featurebyte.api.catalog | activate:157 | Catalog activated: Grocery V1 - playground (spark)
Loading Feature(s) |████████████████████████████████████████| 204/204 [100%] in 1.0s (192.43/s) 
Time taken to create feature payloads: 0:00:01.105876
Time taken to create feature models: 0:00:03.151094
Time taken to create feature list create payload: 0:00:00.155014
Time taken to iterate create feature list create payload: 0:00:00.731023
Time taken to create feature list model: 0:00:01.246440
Time taken to create feature list: 0:00:00.001909
Done! |████████████████████████████████████████| 100% in 0.1s (9.48%/s) 
Loading Feature(s) |████████████████████████████████████████| 204/204 [100%] in 0.9s (210.59/s) 
Time taken to save feature list: 0:00:06.902523
Time taken to delete feature list: 0:00:00.997253
```

<!-- If your PR refers to a related issue, link it here. -->

## Type of Change

<!-- Mark with an `x` all the checkboxes that apply (like `[x]`) -->

Label this pull request to place it under the correct category in Release Notes:

|        **Pull Request Label**         | **Category in Release Notes** |
|:-------------------------------------:|:-----------------------------:|
|       `enhancement`, `feature`        |          🚀 Features          |
| `bug`, `refactoring`, `bugfix`, `fix` |    🔧 Fixes & Refactoring     |
|       `build`, `ci`, `testing`        |    📦 Build System & CI/CD    |
|              `breaking`               |      💥 Breaking Changes      |
|            `documentation`            |       📝 Documentation        |
|            `dependencies`             |    ⬆️ Dependencies updates    |



## Checklist

<!-- Mark with an `x` all the checkboxes that apply (like `[x]`) -->

- [ ] I have read the [`CODE_OF_CONDUCT.md`](https://github.com/featurebyte/featurebyte/blob/main/CODE_OF_CONDUCT.md) and [`CONTRIBUTING.md`](https://github.com/featurebyte/featurebyte/blob/main/CONTRIBUTING.md) guides.
- [ ] I have written tests for the changes made.
- [ ] I have written docstrings in [NumpyDoc format](https://numpydoc.readthedocs.io/en/latest/format.html#docstring-standard)
- [ ] I have labeled my Pull Request correctly